[#1] Extract "not enough data" markers
I changed the `pdftotext` extraction to try to maintain the text's relative
positions, so we can see the structure of the page. It was then a matter of
finding the asterisk closest to the column, if they exist. We did so by
calculating the Euclidian distance between the column names and the potential
asterisks (i.e. the ones that aren't too far away from them).

With this change, we were able to extract all data from Great Britain.